Video Backends: Move backend implementation to new interface.  TextureCacheBase now uses a 'has a' rather than 'is a' relationship for each backend
