+++ Source/Core/VideoBackends/Vulkan/TextureConverter.h
@@ -31,8 +31,7 @@ public:
   bool Initialize();
 
   // Applies palette to dst_entry, using indices from src_entry.
-  void ConvertTexture(TextureCacheBase::TCacheEntry* dst_entry,
-                      TextureCache::TCacheEntry* src_entry, VkRenderPass render_pass,
+  void ConvertTexture(TCacheEntry* dst_entry, TCacheEntry* src_entry, VkRenderPass render_pass,
                       const void* palette, TLUTFormat palette_format);
 
   // Uses an encoding shader to copy src_texture to dest_ptr.
@@ -50,10 +49,10 @@ public:
                                    u32 src_stride, u32 src_height);
 
   bool SupportsTextureDecoding(TextureFormat format, TLUTFormat palette_format);
-  void DecodeTexture(VkCommandBuffer command_buffer, TextureCache::TCacheEntry* entry,
-                     u32 dst_level, const u8* data, size_t data_size, TextureFormat format,
-                     u32 width, u32 height, u32 aligned_width, u32 aligned_height, u32 row_stride,
-                     const u8* palette, TLUTFormat palette_format);
+  void DecodeTexture(VkCommandBuffer command_buffer, TCacheEntry* entry, u32 dst_level,
+                     const u8* data, size_t data_size, TextureFormat format, u32 width, u32 height,
+                     u32 aligned_width, u32 aligned_height, u32 row_stride, const u8* palette,
+                     TLUTFormat palette_format);
 
 private:
   static const u32 ENCODING_TEXTURE_WIDTH = EFB_WIDTH * 4;
@@ -90,7 +89,7 @@ private:
 
   // Returns the command buffer that the texture conversion should occur in for the given texture.
   // This can be the initialization/copy command buffer, or the drawing command buffer.
-  VkCommandBuffer GetCommandBufferForTextureConversion(const TextureCache::TCacheEntry* src_entry);
+  VkCommandBuffer GetCommandBufferForTextureConversion(const TCacheEntry* src_entry);
 
   // Shared between conversion types
   std::unique_ptr<StreamBuffer> m_texel_buffer;
