+++ Source/Core/VideoCommon/VideoCommon.vcxproj.filters
@@ -173,9 +173,6 @@
     <ClCompile Include="AbstractTexture.cpp">
       <Filter>Base</Filter>
     </ClCompile>
-    <ClCompile Include="AbstractRawTexture.cpp">
-      <Filter>Base</Filter>
-    </ClCompile>
     <ClCompile Include="ShaderGenCommon.cpp">
       <Filter>Shader Generators</Filter>
     </ClCompile>
@@ -190,6 +187,12 @@
     </ClCompile>
     <ClCompile Include="UberShaderVertex.cpp">
       <Filter>Shader Generators</Filter>
+    </ClCompile>
+    <ClCompile Include="TextureCacheEntry.cpp">
+      <Filter>Base</Filter>
+    </ClCompile>
+    <ClCompile Include="TextureCacheBackend.cpp">
+      <Filter>Base</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -347,9 +350,6 @@
     <ClInclude Include="AbstractTexture.h">
       <Filter>Base</Filter>
     </ClInclude>
-    <ClInclude Include="AbstractRawTexture.h">
-      <Filter>Base</Filter>
-    </ClInclude>
     <ClInclude Include="AsyncShaderCompiler.h">
       <Filter>Util</Filter>
     </ClInclude>
@@ -361,6 +361,12 @@
     </ClInclude>
     <ClInclude Include="UberShaderVertex.h">
       <Filter>Shader Generators</Filter>
+    </ClInclude>
+    <ClInclude Include="TextureCacheEntry.h">
+      <Filter>Base</Filter>
+    </ClInclude>
+    <ClInclude Include="TextureCacheBackend.h">
+      <Filter>Base</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
